% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arabidopsisThalianaRefseq.R
\docType{data}
\name{arabidopsisThalianaRefseq}
\alias{arabidopsisThalianaRefseq}
\title{Arabidopsis thaliana Refseq Genome}
\format{A DNAStringSet object of length 7.}
\usage{
data(arabidopsisThalianaRefseq)
}
\description{
The Arabidopsis thaliana refseq genome, downloaded using \code{\link[biomartr]{getGenome}}
}
\keyword{datasets}
