% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPackSeqs.R
\name{getPackSeqs}
\alias{getPackSeqs}
\title{Extract Sequences from Ranges}
\usage{
getPackSeqs(packMatches, Genome, output = "DNAStringSet")
}
\arguments{
\item{packMatches}{A dataframe containing genomic ranges and names referring to sequences to be extracted. Obtained from \code{\link{packSearch}}.}

\item{Genome}{A DNAStringSet object containing sequences referred to in \code{packMatches}.}

\item{output}{The type of object to be returned. At default, returns a \code{\link[Biostrings]{DNAStringSet}} object; alternatively, can return a \code{character} vector.}
}
\value{
The transposon sequences extracted from \code{packMatches}.
}
\description{
Gets the sequences referred to in \code{packMatches} and returns the sequences added as an additional column to the dataframe..
}
\author{
Jack Gisby
}
