% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSeqs.R
\name{getSeqs}
\alias{getSeqs}
\title{Get Sequences from Ranges}
\usage{
getSeqs(packMatches, Genome)
}
\arguments{
\item{packMatches}{A dataframe containing genomic ranges and names referring to sequences to be extracted.}

\item{Genome}{A DNAStringSet object containing sequences referred to in \code{packMatches}}
}
\value{
The dataframe \code{packMatches} with an additional \code{seq} feature containing extracted sequences as characters.
}
\description{
Gets the sequences referred to in \code{packMatches} and returns the sequences added as an additional column to the dataframe..
}
\details{
Function intended for internal use.
}
\author{
Jack Gisby
}
