% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTsds.R
\name{getTSDs}
\alias{getTSDs}
\title{Gets the flanking TSD sequences of the TIRs in \code{tirMatches}.}
\usage{
getTSDs(tirMatches, Genome, tsdLength, direction)
}
\arguments{
\item{tirMatches}{A dataframe containing genomic ranges and names referring to TIR sequences.}

\item{Genome}{A DNAStringSet object containing sequences referred to in \code{tirMatches}}

\item{tsdLength}{The length of the TSD region to be retrieved (integer).}

\item{direction}{The direction of the TIR; "+" for forward, "-" for reverse.}
}
\value{
The dataframe \code{tirMatches} with an additional \code{TSD} feature containing flanking TSD sequences as characters.
}
\description{
Gets the flanking TSD sequences of the TIRs in \code{tirMatches}.
}
