% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyPotentialPackElements.R
\name{identifyPotentialPackElements}
\alias{identifyPotentialPackElements}
\title{Primary filtering stage for the \code{packSearch} algorithm. Identifies potential packTYPE transposable elements based on proximity of matching inverted repeats and equality of TSD sequences.}
\usage{
identifyPotentialPackElements(forwardMatches, reverseMatches, Genome,
  elementLength)
}
\arguments{
\item{forwardMatches}{A dataframe containing genomic ranges and names referring to forwards-facing TIR sequences and their respective TSD sequences.}

\item{reverseMatches}{A dataframe containing genomic ranges and names referring to forwards-facing TIR sequences and their respective TSD sequences.}

\item{Genome}{A DNAStringSet object containing the matches referred to in \code{forwardMatches} and \code{reverseMatches}}

\item{element.length}{A vector of two integers containing the minimum and maximum transposable element length.}
}
\value{
A dataframe, \code{packMatches}, containing the locations of potential packTYPE transposable elements in \code{Genome}.
}
\description{
Primary filtering stage for the \code{packSearch} algorithm. Identifies potential packTYPE transposable elements based on proximity of matching inverted repeats and equality of TSD sequences.
}
