% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyTirMatches.R
\name{identifyTirMatches}
\alias{identifyTirMatches}
\title{Searches a DNAStringSet for potential TIRs based on sequence similarity.}
\usage{
identifyTirMatches(tirMatches = NULL, subSeq, Genome, mismatch,
  strand = "*")
}
\arguments{
\item{tirMatches}{A dataframe of previously identified TIRs, in the format generated by \code{GenomicRanges::makeGRangesFromDataFrame}.}

\item{subSeq}{A DNAString object to be searched for.}

\item{Genome}{A DNAStringSet object containing the DNAString objects to be searched.}

\item{mismatch}{The allowable mismatch between \code{subSeq} and a given slice of \code{Genome}. Includes indels. See \code{biostrings::matchPattern}.}
}
\value{
A dataframe, \code{tirMatches}, containing identified matches. Will be appended to \code{tirMatches} if given.
}
\description{
Searches a DNAStringSet for potential TIRs based on sequence similarity.
}
