% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init}
\alias{init}
\title{Initialise Global Variables}
\usage{
init()
}
\description{
Initialise global variables in order to prevent R variable binding issues (e.g. for \code{dplyr} & \code{mapply})
}
\note{
Called upon package build only.
}
\author{
Jack Gisby
}
