% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packClust.R
\name{packClust}
\alias{packClust}
\title{Cluster Transposons using VSEARCH}
\usage{
packClust(packMatches, Genome, identity = 0.6, threads = 1,
  strand = "both", saveFolder,
  vSearchPath = "path/to/vsearch/vsearch-2.14.1-win-x86_64/vsearch.exe")
}
\arguments{
\item{packMatches}{A dataframe of potential transposable elements. Will be saved as a FASTA file for VSEARCH.}

\item{Genome}{The \code{\link[Biostrings]{DNAStringSet}} object used to create the \code{packMatches} dataframe. Used to retrieve sequences for clustering.}

\item{identity}{The sequence identity of two transposable elements in \code{packMatches} required to be grouped into a cluster.}

\item{threads}{The number of threads to be used by VSEARCH.}

\item{strand}{The strand direction (+, - or *) to be clustered.}

\item{saveFolder}{The folder to save output files (uc, blast6out, FASTA)}

\item{vSearchPath}{The location of the VSEARCH executable file.}
}
\value{
Saves cluster information, including a \code{uc} and \code{blast6out} file, to the specified location. Returns the given \code{packMatches} dataframe with an additional column, \code{cluster}, containing cluster IDs.
}
\description{
Cluster potential pack-TYPE elements by sequence similarity.
}
\note{
In order to cluster sequences using VSEARCH, the executable file must first be installed.
}
\references{
VSEARCH may be downloaded from \url{https://github.com/torognes/vsearch}. See \url{https://www.ncbi.nlm.nih.gov/pubmed/27781170} for further information.
}
\author{
Jack Gisby
}
