% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packSearch.R
\name{packSearch}
\alias{packSearch}
\title{General use pipeline function for the Pack-TYPE transposon finding algorithm.}
\usage{
packSearch(subSeq, Genome, mismatch = 0, elementLength, tsdLength)
}
\arguments{
\item{subSeq}{A DNAString object containing the TIR sequence to be searched for.}

\item{Genome}{A DNAStringSet object to be searched.}

\item{mismatch}{The maximum edit distance to be considered for TIR matches (indels + substitions). See \code{biostrings::matchPattern}.}

\item{elementLength}{The maximum element length to be considered, as a vector of two integers. E.g. \code{c(300, 3500)}}
}
\value{
A dataframe, \code{packMatches}, containing elements identified by the algorithm. These may be autonomous or pack-TYPE elements.
}
\description{
General use pipeline function for the Pack-TYPE transposon finding algorithm.
}
