% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBlast6Out.R
\name{readBlast6Out}
\alias{readBlast6Out}
\title{Convert .blast6out Files to Dataframe}
\usage{
readBlast6Out(file)
}
\arguments{
\item{file}{The file path of the .blast6out file.}
}
\value{
A dataframe containing the converted .blast6out file. 
    The file contains the following features:
    \itemize{
        \item Query sequence ID
        \item Target sequence ID
        \item Percenty sequence identity
        \item Alignment length
        \item Number of mismatches
        \item Number of gaps
        \item Base position of alignment start 
            in query sequence
        \item Base position of alignment end in 
            query sequence
        \item Base position of alignment start 
            in target sequence
        \item Base position of alignment end in 
            target sequence
        \item E-value
        \item Bit score
}
}
\description{
Reads .blast6out files (NCBI Blast Format) 
    generated by the VSEARCH
    clustering and alignment algorithms.
}
\details{
The blast6out file is tab-separated text file 
    compatible with NCBI BLAST m8 and NCBI BLAST+ 
    outfmt 6 formats. One cluster/alignment can be 
    found for each line.
}
\examples{
    readBlast6Out(system.file(
        "extdata", 
        "packmatches.blast6out", 
        package = "packFinder"
    ))

}
\references{
VSEARCH may be downloaded from 
\url{https://github.com/torognes/vsearch}. See 
\url{https://www.ncbi.nlm.nih.gov/pubmed/27781170} 
for further information.
}
\seealso{
code{\link{tirClust}}, code{\link{packAlign}}, 
    code{\link{readUc}}, code{\link{packClust}}
}
\author{
Jack Gisby
}
