% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readUC.R
\name{readUc}
\alias{readUc}
\title{Convert .uc Files to Dataframe}
\usage{
readUc(savePath)
}
\arguments{
\item{savePath}{The file path of the .uc file.}
}
\value{
A dataframe, \code{packClusts}, containing the converted .uc file.
}
\description{
Reads .uc files (USEARCH Cluster Format) generated by the VSEARCH clustering algorithm.
}
\author{
Jack Gisby
}
