% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTsds.R
\name{getTsds}
\alias{getTsds}
\title{Get Flanking Terminal Site Duplications}
\usage{
getTsds(tirMatches, Genome, tsdLength, strand)
}
\arguments{
\item{tirMatches}{A dataframe containing genomic ranges and names referring to TIR sequences.}

\item{Genome}{A DNAStringSet object containing sequences referred to in \code{tirMatches}.}

\item{tsdLength}{The length of the TSD region to be retrieved (integer).}

\item{strand}{The strand of the TIR; "+" for forward, "-" for reverse.}
}
\value{
The dataframe \code{tirMatches} with an additional \code{TSD} feature containing flanking TSD sequences as characters.
}
\description{
Gets the flanking TSD sequences of the TIRs in \code{tirMatches}.
}
\details{
Called by \code{\link{packSearch}}. It is recommended to use the general pipeline function \code{\link{packSearch}} for identification of potential pack elements, which returns TSD sequences as a feature of results, however each stage may be called individually.
}
\author{
Jack Gisby
}
