% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyTirMatches.R
\name{identifyTirMatches}
\alias{identifyTirMatches}
\title{Identify Terminal Inverted Repeat Matches}
\usage{
identifyTirMatches(tirMatches = NULL, subSeq, Genome, mismatch = 0,
  strand = "*")
}
\arguments{
\item{tirMatches}{A dataframe of previously identified TIRs, in the format generated by \code{\link{GenomicRanges::makeGRangesFromDataFrame}}.}

\item{subSeq}{A DNAString object to be searched for.}

\item{Genome}{A DNAStringSet object containing the DNAString objects to be searched.}

\item{mismatch}{The allowable mismatch between \code{subSeq} and a given slice of \code{Genome}. Includes indels. See \code{\link{Biostrings::matchPattern}}.}

\item{strand}{The directionality of the search string ("+", "-" or "*").}
}
\value{
A dataframe, \code{tirMatches}, containing identified matches. Will be appended to \code{tirMatches} if given.
}
\description{
Searches a DNAStringSet for potential TIRs based on sequence similarity.
}
\details{
Called by \code{\link{packSearch}}. Function intended for internal use; for non-internal use, use \code{\link{Biostrings::matchPattern}}.
}
\author{
Jack Gisby
}
