% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyTirMatches.R
\name{identifyTirMatches}
\alias{identifyTirMatches}
\title{Identify Terminal Inverted Repeat Matches}
\usage{
identifyTirMatches(subSeq, Genome, mismatch = 0, strand = "*")
}
\arguments{
\item{subSeq}{A \code{\link[Biostrings]{DNAString}} object to be searched for.}

\item{Genome}{A \code{\link[Biostrings]{DNAStringSet}} object containing the DNAString objects to be searched.}

\item{mismatch}{The allowable mismatch between \code{subSeq} and a given slice of \code{Genome}. Includes indels.}

\item{strand}{The directionality of the search string ("+", "-" or "*").}

\item{tirMatches}{A dataframe of previously identified TIRs, in the format generated by \code{\link[GenomicRanges]{makeGRangesFromDataFrame}}.}
}
\value{
A dataframe, \code{tirMatches}, containing identified matches. Will be appended to \code{tirMatches} if given.
}
\description{
Searches a DNAStringSet for potential TIRs based on sequence similarity.
}
\details{
Called by \code{\link{packSearch}}. Used by \code{\link{packSearch}} as an initial filtering stage. \code{\link[Biostrings]{matchPattern}} from Biostrings is used for pattern matching. It is recommended to use the general pipeline function \code{\link{packSearch}} for identification of potential pack elements, however each stage may be called individually.
}
\author{
Jack Gisby
}
