% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readUC.R
\name{readUc}
\alias{readUc}
\title{Function for the conversion of uc files to a dataframe.}
\usage{
readUc(savePath)
}
\arguments{
\item{savePath}{The file path of the .uc file.}
}
\value{
A dataframe, \code{packClusts}, containing the converted .uc file.
}
\description{
Function for the conversion of uc files to a dataframe.
}
