% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tirClust.R
\name{tirClust}
\alias{tirClust}
\title{Analysis of TIR sequence by cluster.}
\usage{
tirClust(packMatches, plot = TRUE, plotSavePath = NULL, k = 5,
  tirLength = 25)
}
\arguments{
\item{packMatches}{Dataframe containing potential packTYPE elements and cluster information.}

\item{plot}{Argument specifying whether the TIR consensus sequences should be plottted as a dendrogram.}

\item{plotSavePath}{File path for the dendrogram plot. If unspecified, the dendrogram plot is not saved.}

\item{k}{The k-mer size to be used for calculating a distance matrix between TIR consensus sequences. See \code{kdistance::kmer}.}

\item{tirLength}{The TIR size to be considered. Consensus sequences will be generated based on the first and last \code{tirLength} bases.}
}
\value{
A list of consensus sequences for each cluster specified in \code{packMatches}.
}
\description{
Analysis of TIR sequence by cluster.
}
